Remove unnecessary Prolog-to-Java conversion from submit filters
Instead of converting intermediate filtering results from Prolog to Java
(and back) in each filtering iteration, perform the conversion only
after the last submit filter is done.

Also change the return type of the SubmitRuleEvaluator.evaluate to
return ListTerm and let the caller decide how to convert it to a Java
list. This is necessary because the ListTerm.toJava method can produce
either List<Term> or List<String> as its result depending on what is
contained in the ListTerm instance.

Change-Id: I70ca979785c766653281994dd3097fc8795e1a87
Signed-off-by: Sasa Zivkov <sasa.zivkov@sap.com>

Former-commit-id: 162c801f4baaaff2b6d089e8fe950a1f3bc8af3d