Fail on LOCK_FAILURE during merge even if we plan to retry
Not failing the current merge attempt on LOCK_FAILURE means the change
state will get updated even though the merge commit failed. This is
bad.

I suspect this was the original intent of leaving the throw outside
the inner switch block in I9a16e969, but of course that was a bug.
Id2771bad went too far in fixing this bug. I think this one gets it
right.

Change-Id: I91c05ed9a2d826a650cb1bed07dec3ba90c0f117

Former-commit-id: eb05714f884d0754a7a84156be415d84f78804ec