Flush persistent H2 cache if the existing entries are incompatible
Some caches (e.g. 'diff_intraline') are peristent so that the entries
survive a Gerrit restart (e.g. because the entries are expensive to
compute). To persist the cache entries Java serialization is used. If
now a class for a cache entry is changed and its serialVersionUID is
updated, already peristent cache entries cannot be deserialized anymore
since their class is incompatible with the new cache entry class.

Flush the cache if incompatible cache entries are detected.

E.g. commit 0632922f5de5cb85515f146d4bbfd24463341c46 [formerly 2424a8a9d595c87106db51da5867ed417eb8a2f5] changed a cache
entry class and as result the existing entries of the 'diff_intraline'
cache could not be loaded anymore, which caused an exception on each
Gerrit startup and when trying to access one of these cached entries.
To get rid of these exeptions the Gerrit administrator had to manually
flush the cache.

Change-Id: Ia158855d983d51335ae2a4df2b73d6c14fc612f0
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 3f461de64e73917ef814a407bc5c4a7c35127e67