Permit booleans in project.config to be inherited
Allow the parent project to supply the value for properties:

  * use content merge
  * require contributor agreement
  * require change id
  * require signed off by

This makes it easier to configure a large number of projects, as
the properties can be defined once on the parent and inherited.

This commit modifies the default behavior when the variable is not
defined in project.config, it now inherits from the parent rather
than being false.

Change-Id: Ic4a6ea6ff23ad16f1fc82b70cc89a480f59db010

Former-commit-id: 2462ccb8df062c512d1676c7d01f82489807b09d