Don't suggest site init if schema version is newer than expected
In the current implementation the schema version check simply
checks if the current version is not the same as expected, and
then raises an error with the message "Run init to upgrade".

However, re-initialising the site will only work if the current
schema version is older than expected.  If the user has tried to
start an older version of Gerrit with a newer test site (i.e. a
site that has already been initialised with a newer version of
Gerrit), re-initialising the site will not work because schema
downgrade is not supported.

Explicitly check if the current version is older or newer than
expected, and give an appropriate error message.

Also rename some variables to make the meaning a bit clearer.

Change-Id: I5dfc8f286615d1d25831b949f9697e42d6c8bef0

Former-commit-id: fa86d9c7dff576311406f293415ee7c63c9cffdc