Use absolute links to Gerrit documentation in Gerrit 2.5 release notes
The relative links to the Gerrit documentation ('../Documentation/') do
not work when the release notes are uploaded to the Gerrit project
homepage, since the 2.5 documentation is there available under
('../Documentation/2.5/'). Having the links as '../Documentation/2.5/'
results in broken links when the release notes are locally generated.

To fix this we now use absolute links
('http://gerrit-documentation.googlecode.com/svn/Documentation/2.5/')
to the Gerrit documentation. This is not ideal since locally it would
be preferred to have links to the local documentation, but at least
there are now no broken links in both scenarios. Other release notes
that link to the Gerrit documentation also use absolute links (e.g. the
release notes for 2.0.21).

Change-Id: I03da90bd56adc69b473185d4818bf640d8ffdb3c
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: d9b9a36e096d7677f590bbf1b2a4409707bcfbdc