Fix for CGit links
The default settings for cgit is to expect url to repositories
that ends with .git and this is not supported by Gerrit. From
the CGit config help file:

remove-suffix::
  If set to "1" and scan-path is enabled, if any repositories
  are found with a suffix of ".git", this suffix will be
  removed for the url and name. Default value: "0". See also:
  scan-path.

This is of importantance, because if you run cgit and have
remove-suffix enabled, it cannot serve certain gits where the
name without .git matches another path. Example:

  <gits_root>/build.git
  <gits_root>/build/scripts.git

Since <gits_root>/build.git will be served as:

  http://server/build/

CGit cannot any longer serve:

  http://server/build/scripts/

...since it parses the git name from left to right until it
has a match with a matching repository.

Running cgit with remove-suffix:0 will work fine though, since
the two paths will be served as:

  http://server/build.git/
  http://server/build/scripts.git/

Hence, most Gerrit/cgit users with complexer git setups must
run cgit with the default setting of not removing the .git
suffix and will probably expect Gerrit to support this setup.

Change-Id: Ib05e576c5f2dc1def37404a9023477135a1324a3

Former-commit-id: 6a4808e5242ae8266018433709ad5e2adf928396