Do not log RepositoryNotFoundException if non-existing project is accessed
Trying to retrieve a non-existing project from the project cache causes
a warning in the error log which prints the full stacktrace of the
corresponding RepositoryNotFoundException.

This is bad because every try to clone a non-existing project or
providing a non-existing project as project argument to any SSH command
triggers this warning in the error log. This is not a server error but
just a bad request from a user, which shouldn't be logged.

In any case the project cache still returns null for non-existing
projects. All callers handle the null case, those that require the
project to be found throw a NoSuchProjectException so that in this case
there is still an exception that can be logged.

Change-Id: Ie606b1af6756036d41387c8a9fe3b29f4c6345b3
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 70fef54cc98139348d1b47cc05574069e635e8e0