Set change state to NEW if merge fails due to non-existing dest branch
If a submitted change fails to merge because the destination branch
doesn't exist anymore, it stays in state 'Submitted, Merge Pending'.
This means Gerrit will re-attempt to merge this change (e.g. on
startup), but this doesn't make sense. Either the branch still doesn't
exist (then there is no need to try merging it) or a new branch with
the old name was created (then it's questionable if the change should
still be merged into this branch). This is why it's better to set the
change back to the 'Review in Progress' state and update it with a
message saying that it couldn't be merged because the destination
branch doesn't exist anymore.

In addition Gerrit currently writes an error into the error log if a
change can't be merged because the destination branch is missing.
That's not really a server error and shouldn't appear in the log. It's
sufficient to have a message on the change.

Change-Id: Ia9089d2d4166d2a442fe5c5a04e6de6913bec952
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 90e2711060b1627ab22838aead3f5c578cd6b9f6