When a file is renamed, include old file name in Patch attribute
When a file is renamed, the value of `file` in the Patch attribute
is the new file name.  The old file name is not included.

This means it is not possible to find the old file name by running
a gerrit query command.

Add a new field `fileOld`. When the file is renamed, this field
contains the old file name.

The name of the existing field is unchanged.  Renaming to `fileNew`
would make the name more meaningful and consistent with `fileOld`,
but would make it incompatible with user scripts that expect it to
be called `file`.

Change-Id: Ie29c9ea31a6d73ab7ff0c91bf0c10dc10c3636d6

Former-commit-id: 877d493809d91bf02f8d3688cc70d1a5350065eb