Clarify error messages related to commit message footer content
Commits can be rejected due to the following issues in the commit
message:

- Change-Id tag is missing or incorrectly formatted
- Multiple Change-Id tags are present
- Signed-off-by tag is missing or does not specify the correct user

However the error messages raised in these cases do not make it clear
that the required tags are expected to be found in the footer (last
paragraph) of the commit message.

Update the error messages and related documentation to explicitly
mention that the tags should be in the commit message footer.

Change-Id: I7d2d2dfb66bed3697745733c4fe19c177e9bc56b

Former-commit-id: a68769eb4f7a1fc1067255ca434dfdb098f1008a