Don't advertise refs/cache-automerge during push
The cache-automerge space is composed of tree objects that were
made on the fly by the server before presenting a merge commit in
the web UI. Clients are unlikely to have these objects locally,
and therefore cannot used them as a preferred base when creating a
pack for transmission to the server. Save a little bit of network
traffic by hiding them.

Change-Id: I806b5b3b91febbd5b63bc2faeef8df4bd6d101d4

Former-commit-id: 3e49de3d1a4321fa1e3ad8068c298d239aab074b