Enable standalone JavaScript plugins
The idea of 'standalone JavaScript plugins' came during the 'JavaScript
plugins' talk at the Gerrit User Summit.

The main concept is to automatically install and expose all *.js files
that are in Gerrit's plugins/ directory.  This should lower the entry
point barrier for JS developers.

There is one limitation of such plugins: they cannot contribute
additional resources.  When such are required, the implementer should
inline them in JS code or use the gerrit-plugin-js-archetype.

Change-Id: Ibad8c3938c0855fcabe6937fd8b4d508a9a2eac3
Signed-off-by: Dariusz Luksza <dariusz@luksza.org>

Former-commit-id: 357a2427085a46c789f0daef782b775007491658