Better review message when a change is rebased in the UI
When a change is rebased in the UI by pressing the rebase button, a
comment is added onto the review.  For example if patch set 1 is the
latest, and the change is rebased, the following message is added:

  Patch Set 1: Rebased

This can be a bit confusing.  This commit changes the message to:

  Patch Set 2: Patch Set 1 was rebased onto the latest head

Change-Id: I22eea934c900eb2561e08d7cc81e174150a2d44c

Former-commit-id: fd72e3178ae5bd0c5c0999a86aaf6af70d24aec2