Add the ability to edit text from ChangeInfoBlock
An edit icon appears in ChangeInfoBlock.  Clicking on it allows the user
to type in a textbox the new topic and add an optional message to be
posted to the change.

Editing the topic is restricted to users who can upload new patchsets.
This seems appropriate since currently a user has to upload a new
patchset to change the topic.

Change-Id: I6a2ceb92f4a40e2ee27a50d045407d8f7921863d

Former-commit-id: 66c01952ef1140f4bda596cc92bc462c93006423