Include authentication scheme in authorization header
This allows the session provider to decide what authorization
scheme the browser should present when it makes JSON REST API
calls to the server. By default we call it Bearer to match an
OAuth2 style of authentication that browsers won't do natively,
but this can be changed by modifying what the session tracking
implementation returns from getAuthorization().

Change-Id: If0bf3d57f5564bd06811b6c1c509b0ad5e525495

Former-commit-id: aada97b178003b55ed65fb4b631b68bae5e8ce11