Only show submit type in change view for open changes
Submit type is per project, not tied to specific changes.

If the project's submit type is changed after a change has been
merged, the displayed information will not be correct; it will
show the current merge type, which is not necessarily the one that
was used at submit time.

If a change has been abandoned, the submit type is not relevant.

Change-Id: I59f4e18a58f57dfea38ece3dccdeb6a8ee944626

Former-commit-id: b9f966f9309eafa3489626bb8db375f1f2fdc943