Only add topic name edit button on open changes
Currently the button is added on all changes regardless of status.

Clicking the button opens the dialog and allows the user to enter
a new topic name and submit, however this results in an internal
server error.

Add a check to prevent the button from being added if the change
is not open.

Change-Id: I48a90f3ee645771945d9911d892acf3f87292a14

Former-commit-id: 920702f8487eea73554c127480570de1f845e986