New Admin command: ls-user-refs
Allows an administrator to query which refs are visible for a
user. The command is helpful for admins when debugging why a
user cannot access certain refs and also to help admins verify
that certain secret refs are not exposed to the wrong groups.

While it is possible today to accomplish this through the suexec
command and some git commands, it is cumbersome and also not all
admins have filesystem access to the server.

Change-Id: I7018906d162e72235b8fd1faf332302d1cee9a18

Former-commit-id: bf53a70b331e25851c468ad6dd3e6fc747a6aa3d