Fix: NPE when editing topic name of change uploaded without topic
If a change is uploaded for review without specifying a topic name,
the topic field in the change record is null.  This causes a Null
Pointer Exception when attempting to compare the old topic name with
the new one.

Add a check to initialise the topic name to empty string if it is
null.

Change-Id: Iae95bf78f87c1af181db327770f74a514258ced5

Former-commit-id: 3d45206d1cbc0f28499e28d5f32a0310a2dbd9d1