Be more liberal about parsing JSON responses
If the response begins with the JSON magic string, remove it before
parsing. If a response is missing this leading string, parse the
response as-is. This opens the door to the server possibly offering
GET responses without this magic header, enabling some forms of
mashups, if the server administrator wants to permit that.

Future commits are still required to make the magic string optional
on responses coming from RestApiServlet, but this prepares the
client for any such future change.

Change-Id: I5d65cd7754304e6e2a3fc2c8483d20a71c17c406

Former-commit-id: 8481916929ab920a1f55bf0514f69833e743d5f1