Set default flag only for the effective default dashboard
When listing the dashboards for a project only the dashboard which is
the effective default of this project should be marked as default. At
the moment any dashboard which is the default dashboard for any
project in the parent line is marked as default. As result for a
project more than one dashboard might be marked as default in the UI.

Change-Id: I99dcb7950e93c7038438449043cffb5eb16dcc1b
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 6594ef212a99a61c0d7c6823029fd17dcd42211d