Fix: Topic name cannot be force-edited on closed change
There is a permission setting to allow a change's topic to be
edited when the change is closed, but this does not work any more
because the refactoring of topic name editing to use the REST API
in change Ia5edbb232bc288acdad8b145956ad275d170629a adds explicit
checks for the change being open before allowing the edit.

Remove the checks to make force-editing work again.

Change-Id: Ib83451c9c7f373ca64a395b804ed337a7026aacd

Former-commit-id: 0e2d568dd69e2fea3dc7b5b9128ce85becd8e187