Shorten search hint to make it fit into the collapsed search box
The old search hint was not fully fitting into the collapsed search box
so that it was partly not displayed. Shorten the search hint to make it
fit. Since we are now having suggestions for the search box it is less
important to cover all cases in the search hint.

Change-Id: I3836085dfaa3c490d5bb4df6d7b2b2f55487ec36
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 5a4bfb6191d94bace42814a5ed4c3c9868e291fc