Fix JSON parser for NativeString in REST API responses
Latest GWT contains a change that prevents the casting of
a JavaScript String object to our NativeString and back
again. Somewhere in chain it throws a ClassCastException and
gives up, preventing the topic editing call to fail in ChangeApi.
This bug doesn't yet show up in the default build of Gerrit, as
the GWT change was made after the 2.5 release.

Switch JSON parsing to the native GWT support path, which also tries
to use the browser's native JSON parser, but can be stricter when
it has to fallback to eval(), thanks to some safety checks the GWT
team added. Once the object is parsed as a JSONValue in GWT-land
we can decide on the correct way to treat it as a JavaScriptObject
for the callback handle.

Change-Id: I027b48b8bb2d3777f0ec2ed33359d374a9997adb

Former-commit-id: 9cb085369646693442feaa5960aadb2ad31e00f7