Fix blinking query string
In 7a53f17a Brad tried to clear the query string when showing a
page that is not a query results page. This makes sense, the query
string doesn't apply in the Groups > List page for example so it
should be empty and displaying the hint text rather than the last
change query the user ran.

Unfortunately the implementation in onLoad() causes the text to
blink in a visible way. Moving this to onShowView() reduces the
update to be within a single JavaScript event loop call, eliminating
the blinking effect.

Change-Id: I1614638a0ef9a4384a83572ba7dc6cc369833176

Former-commit-id: 63e10dbfead35df566ea260936b42195b8e7ad56