Link dashboard title to a url version of itself
When using a stable project dashboard URL, the url
obfuscates the content of the dashboard which can make
it hard to debug a dashboard or copy and modify it.  In
the special case of stable dashboards, make the title a
link to an unstable url version of the dashboard with the
url reflecting the actual dashboard contents the way a
custom dashboard does.

Change-Id: I7916bdd8767b5f0854e4b0e1a05ff8eb3bfc3a16

Former-commit-id: 3decc025e9efb2d545b497f06ac06d9fd388da1b