Cleanup ChangeApi to simplify the code
Reuse a common block of code for creating the RestApi object that will
be used to handle the request. This consolidates the URI formatting
into a single location and makes the code a bit shorter to read.

Fix the formatting on the method Javadocs to be consistent and a
single line of code, these methods aren't complex and only need
one line to describe them.

Only set properties on the Input type if the value != null. This saves
a little bit of JSON formatting to the server. It isn't strictly
necessary, but it makes the example of how to make the REST API calls
as produced by the web UI more clear for tool authors that are trying
to perform the same action and learning by watching the requests.

Change-Id: Icac32a8a81399bd9c0a4d441795bb71c753917a2

Former-commit-id: d12f3e664a69423ca4bc8dfa6a8ec6ba71484eed