Make sure only Gerrit admins can change the parent of a project
Only Gerrit administrators should be able to change the parent of a
project because by changing the parent project access rights and BLOCK
rules which are configured on a parent project can be avoided.

The set-project-parent SSH command already verifies that the caller is
a Gerrit administrator, however project owners can change the parent
project by modifying the project.config file and pushing to the
refs/meta/config branch.

This change ensures that changes to the project.config file that change
the parent project can only be pushed/submitted by Gerrit
administrators.

In addition it is now not possible anymore to
- set a non-existing project as parent (as this would make the project
  be orphaned)
- set a parent project for the All-Projects root project (the root
  project by definition has no parent)
by pushing changes of the project.config file to refs/meta/config.

Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: f7222cbed98deee86d848bed0be4db70b5bd254e