Recommend rebase on Path Conflict
The current message on path conflict says that the author should
"merge (or rebase) the change".

Because of this many users that are new to Git and Gerrit do a
'git merge' when they get this message. As result they end up with two
commits (the commit that implements the change + the merge commit).
After push there is now a second change for the merge commit.

However the intention of the author was to rebase the change and upload
the rebased commit as a new patch set to the change.

This is why the message on path conflict should clearly recommend to do
a 'git rebase' and not a 'git merge' to resolve the conflict.

Change-Id: I340fd22ffe5be893e57a5fb2430f2c590fc328b9
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 51a81aac22f1311d2ec37db1385ae8ced2e805c7