Work around failed database updates during patch set replace
If the database doesn't update when adding a new patch set to a change
the next attempt will fail because the refs/changes/ reference already
exists. Increment to the next patch set id.

This works around the error seen by Bruce at the hackathon, where he
couldn't upload a new patch set to a change on gerrit-review without
intervention from a Googler mucking with the backend database.  The
bug is also possible in other Gerrit installations, but less likely as
the database is usually more reliable.

Change-Id: I8546524d3fdcac827b3dcab0338077f4f4a30205

Former-commit-id: 2cc3ddd2418b152385cdb8031c4493718b8b29b3