Remove PluginLoader from ReceiveCommits
This class should never be injected into anything except plugin
management code (e.g. install or disable a plugin). If a plugin wants
to expose its name in the validation message it should take an
injection for @PluginName String in its validation constructor and
include this into the message string.

If this should be standard behavior for a validation listener,
the listener should be a class the plugin extends that declares
accepting the injection by default:

  public abstract class CommitValidationListener {
    private String name;

	@Inject
	void setPluginName(@PluginName String name) {
      this.name = name;
    }

    public final Result validate(Event ...) {
      return new Result(validateEvent(event) + "(plugin "+name));
    }

    public abstract Result validateEvent(Event ...);
  }

It probably would have also made more sense to declare the listener as
throwing exceptions when it doesn't like a commit, and returning void
if it is acceptable. This simplifies the control flow within plugins,
as the implementor can trivially abort as soon as they find a
condition they don't like.

Validators should have been implemented in terms of the extension-api,
perhaps using a thin shim interface that Gerrit implements to provide
access to members of the RevCommit object, thereby isolating most
validation attempts from the specific JGit version in use. An alternate
implementation could use the plugin-api to get the specific RevCommit.

This change is only about removing the broken dependency on the
PluginLoader. Fixing anything else that might be wrong about commit
validation is left as an exercise for the reader.

Change-Id: Ied64e2523acac8435189ec2793d7cea4586961d4

Former-commit-id: 97824585e1c7c9c78608510c1076039ddba5f295